+++ hardhat.config.js
@@ -4,6 +4,7 @@ require('@openzeppelin/hardhat-upgrades');
 require('hardhat-gas-reporter');
 require('hardhat-contract-sizer');
 require('dotenv').config();
+const { updateContract, restoreContract } = require('./scripts/updateContractToSepolia');
 
 const FORK = process.env.FORK || 'mainnet';
 const FORKING_URL = FORK === 'sepolia' ? process.env.SEPOLIA_ALCHEMY_OR_INFURA_URL : process.env.MAINNET_ALCHEMY_OR_INFURA_URL;
@@ -15,22 +16,26 @@ const TOKEN_SUPPLY = 100000000000n;
 task('implementation')
   .addPositionalParam('contractType')
   .setAction(async (args, hre) => {
+    const { ethers, network } = hre;
+    let originalContract;
+    if (network.name === 'sepolia' && args.contractType === 'bridge') originalContract = updateContract();
     await hre.run('compile');
-    const {
-      ethers,
-      network: { name: network }
-    } = hre;
-    const [signer] = await ethers.getSigners();
-    const signerBalance = await ethers.provider.getBalance(signer.address);
-    const contractName = getContractName(args.contractType);
-    console.log(`\nDeploying ${contractName} implementation on ${network} using account ${signer.address}...`);
-    const contractFactory = await ethers.getContractFactory(contractName);
-    const implementation = await contractFactory.deploy();
-    const impAddress = await implementation.getAddress();
-    const cost = ethers.formatEther(signerBalance - (await ethers.provider.getBalance(signer.address)));
-    console.log(`\nDeployed ${contractName} implementation at ${impAddress} for ${cost} ETH\n`);
-    await delay(20);
-    await verify(impAddress);
+
+    try {
+      const [signer] = await ethers.getSigners();
+      const signerBalance = await ethers.provider.getBalance(signer.address);
+      const contractName = getContractName(args.contractType);
+      console.log(`\nDeploying ${contractName} implementation on ${network} using account ${signer.address}...`);
+      const contractFactory = await ethers.getContractFactory(contractName);
+      const implementation = await contractFactory.deploy();
+      const impAddress = await implementation.getAddress();
+      const cost = ethers.formatEther(signerBalance - (await ethers.provider.getBalance(signer.address)));
+      console.log(`\nDeployed ${contractName} implementation at ${impAddress} for ${cost} ETH\n`);
+      await delay(20);
+      await verify(impAddress);
+    } finally {
+      if (originalContract) restoreContract(originalContract);
+    }
   });
 
 task('deploy')
@@ -39,51 +44,56 @@ task('deploy')
   .addOptionalParam('token')
   .addOptionalParam('owner')
   .setAction(async (args, hre) => {
+    const { ethers, upgrades, network } = hre;
+    let originalContract;
+    if (network.name === 'sepolia' && args.contractType === 'bridge') originalContract = updateContract();
     await hre.run('compile');
-    const {
-      ethers,
-      upgrades,
-      network: { name: network }
-    } = hre;
-    const [signer] = await ethers.getSigners();
-    const signerBalance = await ethers.provider.getBalance(signer.address);
-    const contractName = getContractName(args.contractType);
-    const initArgs = getInitArgs(args, network, signer);
 
-    console.log(`\nDeploying ${contractName} on ${network} using account ${signer.address}...`);
-    const contractFactory = await ethers.getContractFactory(contractName);
-    const proxy = await upgrades.deployProxy(contractFactory, initArgs, { kind: 'uups' });
-    const proxyAddress = await proxy.getAddress();
-    const cost = ethers.formatEther(signerBalance - (await ethers.provider.getBalance(signer.address)));
-    console.log(`\nDeployed ${contractName} at ${proxyAddress} for ${cost} ETH\n`);
+    try {
+      const [signer] = await ethers.getSigners();
+      const signerBalance = await ethers.provider.getBalance(signer.address);
+      const contractName = getContractName(args.contractType);
+      const initArgs = getInitArgs(args, network, signer);
 
-    await delay(30);
-    await verify(await upgrades.erc1967.getImplementationAddress(proxyAddress));
-    await verify(proxyAddress);
+      console.log(`\nDeploying ${contractName} on ${network} using account ${signer.address}...`);
+      const contractFactory = await ethers.getContractFactory(contractName);
+      const proxy = await upgrades.deployProxy(contractFactory, initArgs, { kind: 'uups' });
+      const proxyAddress = await proxy.getAddress();
+      const cost = ethers.formatEther(signerBalance - (await ethers.provider.getBalance(signer.address)));
+      console.log(`\nDeployed ${contractName} at ${proxyAddress} for ${cost} ETH\n`);
+
+      await delay(30);
+      await verify(await upgrades.erc1967.getImplementationAddress(proxyAddress));
+      await verify(proxyAddress);
+    } finally {
+      if (originalContract) restoreContract(originalContract);
+    }
   });
 
 task('upgrade')
   .addPositionalParam('contractType')
   .addPositionalParam('proxyAddress')
   .setAction(async (args, hre) => {
+    const { ethers, upgrades, network } = hre;
+    let originalContract;
+    if (network.name === 'sepolia' && args.contractType === 'bridge') originalContract = updateContract();
     await hre.run('compile');
-    const {
-      ethers,
-      upgrades,
-      network: { name: network }
-    } = hre;
-    const [signer] = await ethers.getSigners();
-    const signerBalance = await ethers.provider.getBalance(signer.address);
-    const contractName = getContractName(args.contractType);
+    try {
+      const [signer] = await ethers.getSigners();
+      const signerBalance = await ethers.provider.getBalance(signer.address);
+      const contractName = getContractName(args.contractType);
 
-    console.log(`\nUpgrading ${contractName} on ${network} using account ${signer.address}...`);
-    const contractFactory = await ethers.getContractFactory(contractName);
-    await upgrades.upgradeProxy(args.proxyAddress, contractFactory);
-    const cost = ethers.formatEther(signerBalance - (await ethers.provider.getBalance(signer.address)));
-    console.log(`\nUpgraded ${contractName} at ${args.proxyAddress} for ${cost} ETH\n`);
+      console.log(`\nUpgrading ${contractName} on ${network} using account ${signer.address}...`);
+      const contractFactory = await ethers.getContractFactory(contractName);
+      await upgrades.upgradeProxy(args.proxyAddress, contractFactory);
+      const cost = ethers.formatEther(signerBalance - (await ethers.provider.getBalance(signer.address)));
+      console.log(`\nUpgraded ${contractName} at ${args.proxyAddress} for ${cost} ETH\n`);
 
-    await delay(20);
-    await verify(await upgrades.erc1967.getImplementationAddress(args.proxyAddress));
+      await delay(20);
+      await verify(await upgrades.erc1967.getImplementationAddress(args.proxyAddress));
+    } finally {
+      if (originalContract) restoreContract(originalContract);
+    }
   });
 
 function getInitArgs(args, network, signer) {
