+++ package.json
@@ -6,7 +6,7 @@
     "pretty:fix": "prettier --config ./.prettierrc.json --write \"**/*.{js,json}\"",
     "pretty:sol": "npx prettier --write --plugin=prettier-plugin-solidity contracts/**/*.sol",
     "tests": "npx hardhat test",
-    "tests-s": "node switchToSepolia.js",
+    "tests-s": "node scripts/testAgainstSepolia.js",
     "coverage": "npx hardhat coverage"
   },
   "description": "Ethereum smart contracts for the Truth Network",
